Renders drag element through app drag element renderer
