+++ app/styles/_ui.scss
@@ -84,3 +84,4 @@
 @import 'ui/diff-options';
 @import 'ui/commit-message-avatar';
 @import 'ui/popover';
+@import 'ui/drag-elements';
