+++ app/styles/ui/_drag-elements.scss
@@ -1,0 +1,5 @@
+@import 'drag-elements/cherry-pick-commit';
+
+#dragElement {
+  position: absolute
+}
