+++ app/styles/ui/drag-elements/_cherry-pick-commit.scss
@@ -1,0 +1,72 @@
+#cherry-pick-commit-drag-element {
+  display: block;
+  position: absolute;
+  height: 50px;
+  min-width: 300px;
+  z-index: 1000;
+  padding: 0;
+
+  .commit-box {
+    box-shadow: 2px 1px 1px 0px var(--box-border-color);
+    border-top: 1px solid var(--box-border-color);
+    border-left: 1px solid var(--box-border-color);
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    background-color: var(--background-color);
+
+    .commit {
+      border-bottom: none;
+    }
+
+    .count {
+      background-image: linear-gradient(#d73a49, darken(#d73a49, 5%));
+      color: white;
+      width: 18px;
+      height: 18px;
+      font-weight: 500;
+      font-size: 11px;
+      position: absolute;
+      top: -22px;
+      left: 20px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      text-align: center;
+      border-radius: 100%;
+    }
+  }
+
+  .copy-message-label {
+    position: absolute;
+    bottom: -35px;
+    left: 0px;
+    background-color: var(--background-color);
+    border: 1px solid var(--box-border-color);
+    color: var(--link-button-color);
+    padding: var(--spacing-third) var(--spacing-half);
+    .branch-name {
+      color: var(--text-color);
+      margin-left: var(--spacing-third);
+    }
+    div {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+    }
+  }
+
+  &.multiple-selected::before {
+    content: '';
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: var(--background-color);
+    z-index: -1;
+    left: 7px;
+    top: 5px;
+    box-shadow: 1px 1px 1px 1px var(--box-border-color);
+  }
+}
